+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix inline path type rewrites introducing imports that conflict with existing private bindings
+
 ## [0.17.1] - 2026-07-10
 
 ### Fixed